move offices controller to api/v1
