+++ config/routes.rb
@@ -3,6 +3,7 @@ Rails.application.routes.draw do
     namespace :v1 do
       resources :users
       resources :services
+      resources :offices
     end
   end
   root "users#index"
