+++ db/schema.rb
@@ -26,6 +26,14 @@ ActiveRecord::Schema[7.0].define(version: 2022_11_28_150819) do
     t.datetime "updated_at", null: false
   end
 
+  create_table "peripherals", force: :cascade do |t|
+    t.string "name"
+    t.text "description"
+    t.decimal "price", default: "0.0"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+  end
+
   create_table "services", force: :cascade do |t|
     t.string "name"
     t.text "description"
