+++ db/seeds.rb
@@ -18,3 +18,12 @@ end
 9.times do 
   Office.create(title: Faker::Lorem.words(number: 1), description: Faker::Lorem.sentence, area: Faker::Lorem.words(number: 1), occupancy: Faker::Number.number(digits: 1), images: Faker::LoremFlickr.image(size: "50x50", search_terms: ['office']), basic_price: Faker::Number.decimal(l_digits: 2), address: Faker::Address.full_address)
 end
+
+9.times do 
+  Peripheral.create(name: Faker::Lorem.words(number: 1), description: Faker::Lorem.sentence, price: Faker::Number.decimal(l_digits: 2))
+end
+
+p 'Created 9 users'
+p 'Created 9 services'
+p 'Created 9 offices'
+p 'Created 9 peripherals'
